Infomercial: link to schema
While reading the README, a question that immediately popped into my head was how does Specmonstash know about the relationship constraints. Does it actually need a DB and analyse its metadata? Or does it somehow derive it from Spec?

It took me a long time until I came to the 02. Schema part to learn how that works. For me to understand this library and how I could use it, it is a key part to know how it gains knowledge of the entities & relations so I think it is worth pointing to it early in the infomercial, without overwhelming the reader with too much/detailed information.